+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **New commands:**
     - `fix`: Fix or adjust _Dolby Vision RPU(s)_
+    - `generate`: Generate _Dolby Vision P8 RPU_ for _HDR10_ video(s)
     - `png`: Extract _video frame(s)_ as `PNG` image(s)
     - `mp3`: Extract _audio track(s)_ as `MP3` file(s)
 - **Rewritten info command:**